:truck: refactor: update participant progress tracking handler
